+++ database/migrations/2022_03_23_090828_create_chuc_vus_table.php
@@ -17,7 +17,7 @@ class CreateChucVusTable extends Migration
             $table->id();
             $table->string('kho_id');
             $table->unsignedInteger('ten_chuc_vu');
-            $table->float('luong_co_ban');
+            $table->double('luong_co_ban');
             $table->timestamps();
             $table->softDeletes();
         });
