+++ database/migrations/2022_03_23_091442_create_phu_caps_table.php
@@ -15,7 +15,7 @@ class CreatePhuCapsTable extends Migration
     {
         Schema::create('phu_caps', function (Blueprint $table) {
             $table->id();
-            $table->float('tien_phu_cap');
+            $table->double('tien_phu_cap');
             $table->timestamps();
             $table->softDeletes();
         });
