+++ database/migrations/2022_03_23_092419_create_thuongs_table.php
@@ -16,7 +16,7 @@ class CreateThuongsTable extends Migration
         //https://sites.google.com/site/quanlyluongnhanvien/quan-ly-luong-nhan-vien (cách tính lương)
         Schema::create('thuongs', function (Blueprint $table) {
             $table->id();
-            $table->float('tien_phu_cap');
+            $table->double('tien_phu_cap');
             $table->timestamps();
             $table->softDeletes();
         });
